requirements.txt
streamlit
python-pptx
pypdf2
reportlab
ollama
fitz